requirements.txt
flask
requests
gunicorn
pillow